Crate ticky

Ticky

A simple stopwatch implementation.

Example

use ticky::Stopwatch;
let mut sw = Stopwatch::start_new();
// Do something …
sw.stop();
println!("Elapsed time: {}ms", sw.elapsed_ms_whole());

Features

derive_more - Enables using derive_more for deriving From, Into, Mul, MulAssign, Div, DivAssign, Rem, Shr, and Shl for Stopwatch.

hifitime - Enables using hifitime for high-resolution timekeeping.

stdtime - Enables using std::time for timekeeping.

Either hifitime or stdtime must be enabled. If neither is enabled, stdtime is used by default. If both are enabled, hifitime is used.

Installation

Run cargo add ticky to add Ticky to your Cargo.toml file.

License

Ticky is licensed under the GNU Affero General Public License.

Contributing

Contributions are welcome! Please see CONTRIBUTING.md for more information.

Authors

Emil Sayahi

Acknowledgements

rust-stopwatch - The inspiration for this crate.

Structs

Stopwatch

A simple stopwatch implementation.